+++ minimalkv/__init__.py
@@ -5,12 +5,11 @@ from minimalkv._constants import (
     VALID_KEY_REGEXP,
     VALID_NON_NUM,
 )
-from minimalkv._get_store import get_store, get_store_from_url
+from minimalkv._get_store import get_store_from_url
 from minimalkv._key_value_store import KeyValueStore, UrlKeyValueStore
 from minimalkv._mixins import CopyMixin, TimeToLiveMixin, UrlMixin
 from minimalkv._store_creation import create_store
 from minimalkv._store_decoration import decorate_store
-from minimalkv._urls import url2dict
 
 try:
     import pkg_resources
@@ -25,11 +24,9 @@ __all__ = [
     "decorate_store",
     "FOREVER",
     "get_store_from_url",
-    "get_store",
     "KeyValueStore",
     "NOT_SET",
     "TimeToLiveMixin",
-    "url2dict",
     "UrlKeyValueStore",
     "UrlMixin",
     "VALID_KEY_RE",